git-svn-id: svn://svn.code.sf.net/p/peachfuzz/svn/branches/Peach2.3@1498 bf843fa0-5fda-4187-82bf-2edb81848ee6

+++ Peach/Mutators/number.py
@@ -89,7 +89,7 @@ class NumericalVarianceMutator(Mutator):
 	supportedDataElement = staticmethod(supportedDataElement)
 	
 	def sequencialMutation(self, node):
-		node.currentValue = node.getInternalValue() - self._values[self._currentCount]
+		node.currentValue = long(node.getInternalValue()) - self._values[self._currentCount]
 	
 	def randomMutation(self, node):
 		try:
